updated address fetch function
