+++ client/components/form/Form.jsx
@@ -12,7 +12,6 @@ import {
   Snackbar,
 } from "@mui/material";
 import { Button } from "@mui/material";
-import AddressInput from "./AddressInput";
 import TimeRangeInput from "./TimeRangeInput";
 import { requestAppt } from "@/actions/form";
 import AutocompleteAddress from "./AutocompleteAddress";
