+++ client/constants/index.js
@@ -1,2 +1,3 @@
 // Environment variables
 export const serverUrl = process.env.SERVER_URL;
+export const laCityKey = process.env.NEXT_PUBLIC_LACITY_KEY;
